error 400 for role not found

+++ backend/tracim_backend/lib/core/userworkspace.py
@@ -3,10 +3,12 @@ import typing
 
 from sqlalchemy.orm import Query
 from sqlalchemy.orm import Session
+from sqlalchemy.orm.exc import NoResultFound
 
 from tracim_backend.config import CFG
 from tracim_backend.exceptions import RoleAlreadyExistError
 from tracim_backend.exceptions import UserCantRemoveHisOwnRoleInWorkspace
+from tracim_backend.exceptions import UserRoleNotFound
 from tracim_backend.models.auth import User
 from tracim_backend.models.context_models import UserRoleWorkspaceInContext
 from tracim_backend.models.data import UserRoleInWorkspace
@@ -99,7 +101,17 @@ class RoleApi(object):
             filter(UserRoleInWorkspace.user_id == user_id)
 
     def get_one(self, user_id: int, workspace_id: int) -> UserRoleInWorkspace:
-        return self._get_one_rsc(user_id, workspace_id).one()
+        try:
+            user_role = self._get_one_rsc(user_id, workspace_id).one()
+        except NoResultFound as exc:
+            raise UserRoleNotFound(
+                'Role for user {user_id} '
+                'in workspace {workspace_id} was not found.'.format(
+                    user_id=user_id,
+                    workspace_id=workspace_id,
+                )
+            )
+        return user_role
 
     def update_role(
         self,
@@ -155,6 +167,14 @@ class RoleApi(object):
         if self._user and self._user.user_id == user_id:
             raise UserCantRemoveHisOwnRoleInWorkspace(
                 "user {} can't remove is own role in workspace".format(user_id)
+            )
+        if self._get_one_rsc(user_id, workspace_id).count() == 0:
+            raise UserRoleNotFound(
+                'Role for user {user_id} '
+                'in workspace {workspace_id} was not found.'.format(
+                    user_id=user_id,
+                    workspace_id=workspace_id,
+                )
             )
         self._get_one_rsc(user_id, workspace_id).delete()
         if flush:
